Add fetch file details async job. Reorganized project structure.
