+++ src/.flake8
@@ -1,0 +1,5 @@
+[flake8]
+ignore = E501,E402
+per-file-ignores = 
+    __init__.py: F401
+exclude = .git,__pycache__,docs/source/conf.py,old,build,dist
