+++ src/api/exceptions.py
@@ -1,0 +1,10 @@
+class APIException(Exception):
+    pass
+
+
+class FileInvalidStatusError(Exception):
+    pass
+
+
+class FileNotFound(Exception):
+    pass
