+++ src/api/requirements.txt
@@ -1,3 +1,4 @@
+flake8==3.9.0
 Flask==1.1
 gunicorn==20.0.4
 gevent==20.12.1
