+++ src/api/utils.py
@@ -1,0 +1,4 @@
+class FileStatus(object):
+    PROCESSING = 'PROCESSING'
+    FINISHED = 'FINISHED'
+    FAILED = 'FAILED'
