+++ src/core/logging_setup/__init__.py
@@ -1,1 +1,1 @@
-from core.logging_setup.main import setup_logging
+from core.logging_setup.main import setup_logging  # flak8: noqa
