+++ src/docker-compose.yml
@@ -20,5 +20,5 @@ services:
       - 9002:9001
     environment:
       - FLASK_ENV=development
-      - PROCESSING_API_HOST=http://mock_api:9001
+      - PROCESSING_API_HOST=http://interview-api.snackable.ai
       - MAX_PAGES=200
